feat: Support cross-line routing and custom alert distance in UI
- Flatten available stations and remove 'Metro Line' filter requirement
- Allow users to set custom alert thresholds between 200m and 1000m using a UI slider
- Add @Serializable annotations to domain models to allow Intent passage of routing details
- Wire MultiLineRouteFinder into the ViewModel to assemble full cross-line paths
- Serialize the entire PlannedRoute across the Intent to support TripService's proximity events
- Display relevant interchange transfer hints in Android Notifications

+++ app/src/main/java/com/example/hmrcompanion/domain/TripProgressManager.kt
@@ -2,7 +2,7 @@ package com.example.hmrcompanion.domain
 
 class TripProgressManager(
     private val route: List<Station>,
-    private val alertThresholdMeters: Double = 400.0
+    private val alertThresholdMeters: Int = 400
 ) {
     internal var lastConfirmedIndex: Int? = null
